Create low level formatting module
