+++ src/logging.c
@@ -1,22 +1,11 @@
 #include "logging.h"
 
 /**
- * @brief Make `printf` send characters through the USART.
- *
- * Both standard output and standard error are redirected through the USART.
+ * @brief Log the current battery voltage.
  */
-int _write(int file, char *ptr, int len)
+void log_battery_voltage(void)
 {
-	int i;
-
-	if (file == STDOUT_FILENO || file == STDERR_FILENO) {
-		for (i = 0; i < len; i++)
-			usart_send_blocking(USART3, ptr[i]);
-		return i;
-	}
-
-	errno = EIO;
-	return -1;
+	LOG_INFO("%f", get_battery_voltage());
 }
 
 /**
